Remove unnecessary create_wallet function
In the create_wallet function, we observed an issue related to package dependencies. This issue arises when the utility_helpers package is imported, triggering the installation of the requests library. Unfortunately, requests may not be a pre-installed package in some of our environments, leading to installation failures and subsequent errors